Products terminado

+++ src/controllers/productsController.js
@@ -1,10 +1,11 @@
-import ProductsManagerMongoose from "../daos/mongoose/ProductManagerMongoose.js";
+import ProductsManagerMongoose from "../daos/mongoose/productDao.js";
 
 const productManager = new ProductsManagerMongoose()
 
 export const getAllController = async (req, res, next) => {
     try {
         const docs = await productManager.getAllProducts()
+        console.log('¡Productos obtenidos!')
         res.json(docs)
     } catch (error) {
         next(error)
@@ -18,6 +19,7 @@ export const getProductsByIdController = async (req, res, next) => {
         if (!docs) {
             throw new Error('El producto NO existe.')
         } else {
+            console.log('¡Producto obtenido por ID!')
             res.json(docs)
         }
     } catch (error) {
@@ -40,7 +42,8 @@ export const createProductController = async (req, res, next) => {
         if (!newProduct) {
             throw new Error('No se pudo crear el producto.')
         } else {
-            res.json('Producto creado:', newProduct)
+            console.log('¡Producto creado!')
+            res.json(newProduct)
         }
     } catch (error) {
         next(error)
@@ -51,19 +54,29 @@ export const updateProductController = async (req, res, next) => {
     try {
         const { id } = req.params
         const { title, description, category, code, price, thumbnail, stock } = req.body;
+
+        const docs = await productManager.getProductById(id)
+
+        if(!docs){
+            throw new Error ('El producto NO existe y, por lo tanto, NO puede ser actualizado.')
+        }
+
         const updateProduct = await productManager.updateProduct(id, {
             title,
             description,
-            category, code,
+            category, 
+            code,
             price,
             thumbnail,
             stock
         });
 
         if (!updateProduct) {
-            throw new Error('No se pudo actalizar el producto.')
+            throw new Error('No se pudo actualizar el producto.')
         } else {
-            res.json('Producto actualizado:', updateProduct)
+            const finalProduct = await productManager.getProductById(id)
+            console.log('¡Producto actualizado!')
+            res.json(finalProduct)
         }
 
     } catch (error) {
@@ -71,14 +84,22 @@ export const updateProductController = async (req, res, next) => {
     }
 }
 
-export const deleteProductController = async () => {
+export const deleteProductController = async (req, res, next) => {
     try {
         const { id } = req.params
         
+        const productFounded = await productManager.getProductById(id)
+
+        if(!productFounded){
+            throw new Error ('Producto NO encontrado.')
+        }
+
         const deleteProduct = await productManager.deleteProduct(id)
-        if(!deleteProduct){
-            throw new Error ('No se pudo borrar el producto');
-        }else{
+
+
+        if (!deleteProduct) {
+            throw new Error('No se pudo borrar el producto');
+        } else {
             console.log('¡Producto borrado!')
             res.json(deleteProduct)
         }
